Abstracted String Manipulations
Cleaned up the program a little and put some more general use, not specific to this program, functions away into another file for future use.
